Some more options to use spiderweb graph and polar chart (need to include highcharts-more.js)
 * data-graph-polar
"true" to render a polar graph.
Default: undefined

 * data-graph-yaxis-{X}-grid-line-interpolation
Define the graph type.
Default: null
ex: "polygon" to get a spiderweb graph

 * data-graph-xaxis-show-last-label
If you have a number on the last label, put it to false.
Default: true

 * tickmarkPlacement
For categorized axes only. If "on" the tick mark is placed in the center of the category, if "between" the tick mark is placed between categories.
default: "on"

 * data-graph-xaxis-gridLine-width
The width of the grid lines extending the ticks across the plot area.
Default: 0