fix jsfiddle examples and now they are embeded
the js http://highcharts.com/js/testing.js is not available (highchart change
the way they do demos on jsfiddle).
So now the demos files are included like this in the fiddle :

<script src="http://code.highcharts.com/highcharts.js"></script>
<script src="http://code.highcharttable.org/master/jquery.highchartTable-min.js"></script>

Also, before, the link make the user change page to jsfiddle. Now the fiddle is embeded
in a modal.